pyproject.toml
[project]
name = "cspc-talk"
version = "0.1.0"
description = "This project is the implementation of the solutions for my CSPC talk. It is meant to be illustrative only. Please do not use in a production environment. All data provided is fictional and synthetic."
authors = [
    {name = "John Sixto Santos",email = "johnsixtosantos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "llama-index (>=0.12.29,<0.13.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "llama-index-llms-openai (>=0.3.32,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "plotly (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
